requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
asgiref==3.8.1
attrs==24.2.0
Authlib==1.3.2
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.17.1
azure-mgmt-core==1.5.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.1.1
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
boto3==1.35.56
botocore==1.35.56
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Django==5.1.2
django-filter==24.3
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
eventlet==0.36.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
image==1.5.33
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jmespath==1.0.1
jsonformatter==0.3.2
keyring==25.5.0
ldap3==2.9.1
libgravatar==1.0.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
numpy==2.1.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
passlib==1.7.4
pgadmin4==8.12
pillow==11.0.0
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.2.0
pypng==0.20220715.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.3
setuptools==74.1.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
soupsieve==2.6
speaklater3==1.4
SQLAlchemy==2.0.36
sqlparse==0.5.1
sshtunnel==0.4.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
views==0.3
Werkzeug==3.0.6
wsproto==1.2.0
WTForms==3.1.2
yarl==1.16.0
zstandard==0.23.0
